com 1

+++ db/psql/models.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, Enum
+from sqlalchemy import Column, Integer, String, Enum, BigInteger
 
 from db.psql.connect import Base
 from db.psql.enums.enums import Roles
@@ -8,7 +8,7 @@ class User(Base):
     __tablename__ = "users"
 
     id = Column(Integer, primary_key=True, index=True)
-    tg_id = Column(Integer, nullable=True)
+    tg_id = Column(BigInteger, nullable=True)
     username = Column(String, unique=True, index=True)
     first_name = Column(String, nullable=True)
     last_name = Column(String, nullable=True)
